[COMPLETE] LP-1674 OTHER Send emails to prompt learners to view & download certificates. moving custom signals to philu app.
